Update LICENSE
Today I saw a talk by Peter Leppelt, the founder and CEO of Qabel, and am now very convinced of the open QaPL license. It excludes military and intelligence use. That also corresponds to my conviction. Therefore ClearURLs is immediately below the QaPL 0.2. However, this will not change anything for the users and possible co-developers.