Version 1.3.3.1
I've updated the ClearURLs core, so that ClearURLs only cleans the fields from urls.

This means that ClearURLs can no longer examine all components of a URL and, if necessary, remove something, but only the so-called GET fields.

This restricts not hardly, since everything that is not a GET field is a part of the url (the path) and should not actually be removed.

However, because ClearURLs was able to remove these path elements, this problem came up again. Now ClearURLs can only remove GET fields and thus no longer destroy paths.

I hope that there are fewer problems now. Another little side effect of the change, ClearURLs is now working more efficiently.