New Feature - Report URLs
Lately, I have received relatively few requests with potentially bad URLs. On the one hand, that could be because ClearURLs are working very well by now. But I also think that not everyone wants to sign up to GitLab to report a URL. That's why I've added a new button that reports the current URL in the active tab to me, so I can take a closer look at the URL.

So that you can see exactly what data will be sent to me, I have uploaded the source code for the PHP script, of course, without a password etc.

Please be careful not to send URLs with personal information or tokens.