Make start.getKeys() in clearurls.js ignore method names
start.getKeys() was using a for...in loop to iterate over fields in the
given object, meaning that methods names (e.g. getOrDefault()) get
included in the iteration as well. This causes erroneous entries to be
added the global list of providers. Fix this by using Object.keys() to
extract field names and iterate over that instead.